requirements.txt
alembic==1.12.0
amqp==5.1.1
anyio==3.7.1
async-timeout==4.0.3
billiard==4.1.0
black==23.12.1
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.1.3
fastapi==0.103.1
flower==2.0.1
greenlet==2.0.2
h11==0.14.0
humanize==4.8.0
idna==3.4
Jinja2==3.1.2
kombu==5.3.2
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pydantic==1.10.12
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.6
